joystick + mvvm prep

+++ PTM_PROJECT/src/Model_Pack/Model.java
@@ -22,7 +22,7 @@ public class Model extends Observable {
 		i.interpet();
 	}
 
-	public int normalize(int x, int max, int min, int newmax, int newmin) {
+	public Double normalize(Double x, Double max, Double min, Double newmax, Double newmin) {
 		return (x - min) / (max - min) * (newmax - newmin) + newmin;
 	}
